+++ apps/registration/.eslintrc.js
@@ -1,8 +1,0 @@
-module.exports = {
-  root: true,
-  extends: '@teo0098/eslint-config',
-  parserOptions: {
-    tsconfigRootDir: __dirname,
-  },
-  ignorePatterns: ['.eslint*', '*config*'],
-};