+++ libs/ui/src/components/Box/index.ts
@@ -1,0 +1,2 @@
+export * from './Box.component';
+export * from './Box.types';