+++ libs/ui/src/components/PasswordInput/index.ts
@@ -1,0 +1,3 @@
+export * from './PasswordInput.component';
+export * from './PasswordInput.constants';
+export * from './PasswordInput.types';