+++ libs/ui/src/components/Popover/index.ts
@@ -1,0 +1,2 @@
+export * from './Popover.component';
+export * from './Popover.types';